+++ src/KeyFrame.cc
@@ -143,7 +143,7 @@ Sophus::SE3f KeyFrame::GetPoseInverse()
     return mTwc;
 }
 
-void KeyFrame::GetImgsAndPose(uint8_t & id, cv::Mat & imRGB, cv::Mat & imDepth, Sophus::SE3f & pose)
+void KeyFrame::GetImgsAndPose(uint64_t & id, cv::Mat & imRGB, cv::Mat & imDepth, Sophus::SE3f & pose)
 {
     unique_lock<mutex> lock(mMutexPose);
     id = mnId;
